Add diff-api-snapshot action to danger (#52045)
Summary:
This is a work in progress PR that adds breaking change detection to danger.

The danger is ran on a base revision from which there is no easy access to the API snapshot from the PR. The Github action fetches the PR and stores the "current" snapshot in the temporary directory which is then used in the `diff-api-snapshot` script. It is compared to the previous snapshot from the base revision which can be easily accessed by reading from `packages/react-native/ReactNativeApi.d.ts`.


## Changelog:

<!-- Help reviewers and the release process by writing your own changelog entry.

Pick one each for the category and type tags:

[ANDROID|GENERAL|IOS|INTERNAL] [BREAKING|ADDED|CHANGED|DEPRECATED|REMOVED|FIXED|SECURITY] - Message

For more details, see:
https://reactnative.dev/contributing/changelogs-in-pull-requests


Test Plan:
Tested on react-native fork with personal access token. 

 {F1979339175} 

https://github.com/coado/react-native/pull/16

Differential Revision: D76735630

Pulled By: coado